Remove IntervalPatternTokeniser
The interval pattern tokeniser allowed for the creation of strange tokens
with a mix of separate/join decisions incompatible with the training corpus.
Removing it means we now merged the CompoundPatternTokeniser and
PatternTokeniser classes.
