+++ talismane_core/src/main/resources/reference.conf
@@ -213,16 +213,10 @@ talismane {
 			
 			# If using a pattern tokeniser, the path to the tokeniser model
 			#model = null
-			
-			### If using a pattern tokeniser, the method: Interval or Compound
-			pattern-tokeniser-type = "Compound"
 						
 			# The beam width used by the tokeniser
 			beam-width = 1
 						
-			# if true, will spew out a file full of low-level analysis details for certain model types
-			include-details = false
-			
 			output = ${talismane.core.output} {
 				# if provided, overrides the builtInTemplate by a specific a FreeMarker template for writing the output
 				#template = null
